Change ushort to uint to remain compatiable with standard math functions.

+++ matLib/matComplex.c
@@ -14,11 +14,11 @@
 
 
 /*******************************************************************************
-* matrixz* mat_initz ( ushort rows, ushort cols )
+* matrixz* mat_initz ( uint rows, uint cols )
 * Initializes a new matrix with the specified dimensions,
 * and sets the elements to complex values of zero.
 *******************************************************************************/
-// matrixz* mat_initz ( ushort rows, ushort cols ) {
+// matrixz* mat_initz ( uint rows, uint cols ) {
 //   mat_err( ( rows<1 || cols<1 ), "Error (mat_initz): Matrix dimensions must be positive." );
 //   matrixz *out;
 //   out = (matrixz *) malloc( sizeof(matrixz) );
@@ -84,7 +84,7 @@
 * Display a complex matrix in the terminal.
 *******************************************************************************/
 // void mat_printz ( matrixz* mat ) {
-//   ushort r, c, i, j;
+//   uint r, c, i, j;
 //   char sign;
 //   float re, im;
 //   r = mat->rows;
@@ -104,8 +104,8 @@
 // }
 /*
   printf( "[%dx%d]\n", mat->rows, mat->cols );
-  for( ushort i=1; i<=mat->rows; i++ ) {
-    for( ushort j=1; j<=mat->cols; j++ ) {
+  for( uint i=1; i<=mat->rows; i++ ) {
+    for( uint j=1; j<=mat->cols; j++ ) {
       re = mat_getre( mat, i, j );
       im = mat_getim( mat, i, j );
       if(im<0)  sign = '-';
@@ -125,7 +125,7 @@
 *******************************************************************************/
 // matrixz* mat_readz ( char* file ) {
 //   FILE *f;
-//   ushort i, r, c, n;
+//   uint i, r, c, n;
 //   int scan;
 //   float re, im;
 //   char sign;
@@ -155,7 +155,7 @@
 //   return out;
 // }
 /*
-  ushort r, c;
+  uint r, c;
   int scan;
   float val;
 
@@ -171,11 +171,11 @@
   scan = fscanf( f, "%d", &c );
   mat_err( ( scan == EOF ), "Error (mat_readz): Failed to read 'cols' from file." );
 
-  ushort n = r * c;
+  uint n = r * c;
   matrixz* out = mat_init( r, c );
   float complex* data = out->data;
 
-  for( ushort i=0; i<n; i++ ) {
+  for( uint i=0; i<n; i++ ) {
     //scan = fscanf( f, "%f", &val );
     scan = fscanf( f, "%f %c %f i", &re, &sign, &im );
     mat_err( ( scan == EOF ), "Error (mat_readz): Matrix is missing elements." );
@@ -201,7 +201,7 @@
 *******************************************************************************/
 // void mat_writez ( matrixz* mat, char* file ) {
 //   FILE *f;
-//   ushort r, c, i, j;
+//   uint r, c, i, j;
 //   float re, im;
 //   char sign;
 //   r = mat->rows;
@@ -227,8 +227,8 @@
   mat_err( ( f == NULL ), "Error (mat_writez): Cannot open file.\n" );
 
   fprintf( f, "# %d %d \n", mat->rows, mat->cols );
-  for( ushort i=1; i<=mat->rows; i++ ) {
-    for( ushort j=1; j<=mat->cols; j++ ) {
+  for( uint i=1; i<=mat->rows; i++ ) {
+    for( uint j=1; j<=mat->cols; j++ ) {
       float re = mat_getre( mat, i, j );
       float im = mat_getim( mat, i, j );
       if(im<0)  sign = '-';
@@ -245,17 +245,17 @@
 
 
 /*******************************************************************************
-* float mat_getre ( matrixz* mat, ushort row, ushort col )
+* float mat_getre ( matrixz* mat, uint row, uint col )
 * Returns the real part of a complex matrix element.
 *******************************************************************************/
-// float mat_getre ( matrixz* mat, ushort row, ushort col ) {
+// float mat_getre ( matrixz* mat, uint row, uint col ) {
 //   mat_err( row > mat->rows, "Error (mat_getre): Row index exceeds matrix dimensions."    );
 //   mat_err( col > mat->cols, "Error (mat_getre): Column index exceeds matrix dimensions." );
 //   mat_err( row < 1,         "Error (mat_getre): Row index must be positive."             );
 //   mat_err( col < 1,         "Error (mat_getre): Column index must be positive."          );
 //   float re;
 //   float complex *data = mat->data;
-//   ushort offset = (row-1) * (mat->cols) + (col-1);
+//   uint offset = (row-1) * (mat->cols) + (col-1);
 //   data += offset;
 //   re = creal(*data);
 //   return re;
@@ -271,17 +271,17 @@
 
 
 /*******************************************************************************
-* float mat_getim ( matrixz* mat, ushort row, ushort col )
+* float mat_getim ( matrixz* mat, uint row, uint col )
 * Returns the imaginary part of a complex matrix element.
 *******************************************************************************/
-// float mat_getim ( matrixz* mat, ushort row, ushort col ) {
+// float mat_getim ( matrixz* mat, uint row, uint col ) {
 //   mat_err( row > mat->rows, "Error (mat_getim): Row index exceeds matrix dimensions."    );
 //   mat_err( col > mat->cols, "Error (mat_getim): Column index exceeds matrix dimensions." );
 //   mat_err( row < 1,         "Error (mat_getim): Row index must be positive."             );
 //   mat_err( col < 1,         "Error (mat_getim): Column index must be positive."          );
 //   float im;
 //   float complex *data = mat->data;
-//   ushort offset = (row-1) * (mat->cols) + (col-1);
+//   uint offset = (row-1) * (mat->cols) + (col-1);
 //   data += offset;
 //   im = cimag(*data);
 //   return im;
@@ -297,13 +297,13 @@
 
 
 /*******************************************************************************
-* matrixz* mat_getrz ( matrixz* mat, ushort row ) {
+* matrixz* mat_getrz ( matrixz* mat, uint row ) {
 * Returns the specified complex row vector of a matrix.
 *******************************************************************************/
-// matrixz* mat_getrz ( matrixz* mat, ushort row ) {
+// matrixz* mat_getrz ( matrixz* mat, uint row ) {
 //   mat_err( row > mat->rows, "Error (mat_getrz): Row index exceeds matrix dimensions." );
 //   mat_err( row < 1,         "Error (mat_getrz): Row index must be positive."          );
-//   ushort i, c;
+//   uint i, c;
 //   matrixz *out;
 //   c = mat->cols;
 //   out = mat_initz(1,c);
@@ -315,7 +315,7 @@
 
   matrixz* out = mat_initz( 1, mat->cols );
 
-  for( ushort i=0; i<mat->cols; i++ )  *(out->data+i) = *( mat->data + (row-1) * mat->cols + i );
+  for( uint i=0; i<mat->cols; i++ )  *(out->data+i) = *( mat->data + (row-1) * mat->cols + i );
 
   return out;
 */
@@ -324,13 +324,13 @@
 
 
 /*******************************************************************************
-* matrixz* mat_getcz ( matrixz* mat, ushort col )
+* matrixz* mat_getcz ( matrixz* mat, uint col )
 * Returns the specified complex column vector of a matrix.
 *******************************************************************************/
-// matrixz* mat_getcz ( matrixz* mat, ushort col ) {
+// matrixz* mat_getcz ( matrixz* mat, uint col ) {
 //   mat_err( col > mat->cols, "Error (mat_getcz): Column index exceeds matrix dimensions." );
 //   mat_err( col < 1,         "Error (mat_getcz): Column index must be positive."          );
-//   ushort i, r;
+//   uint i, r;
 //   matrixz *out;
 //   r = mat->rows;
 //   out = mat_initz(r,1);  
@@ -342,7 +342,7 @@
 
   matrixz* out = mat_initz( mat->rows, 1 );
 
-  for( ushort i=0; i<mat->rows; i++ )  *(out->data+i) = *( mat->data + ( i * mat->cols ) + (col-1) );
+  for( uint i=0; i<mat->rows; i++ )  *(out->data+i) = *( mat->data + ( i * mat->cols ) + (col-1) );
 
   return out;
 */
@@ -351,16 +351,16 @@
 
 
 /*******************************************************************************
-* void mat_setz ( matrixz* mat, ushort row, ushort col, float re, float im )
+* void mat_setz ( matrixz* mat, uint row, uint col, float re, float im )
 * Assigns a complex number into a matrix element.
 *******************************************************************************/
-// void mat_setz ( matrixz* mat, ushort row, ushort col, float re, float im ) {
+// void mat_setz ( matrixz* mat, uint row, uint col, float re, float im ) {
 //   mat_err( row > mat->rows, "Error (mat_setz): Row index exceeds matrix dimensions."    );
 //   mat_err( col > mat->cols, "Error (mat_setz): Column index exceeds matrix dimensions." );
 //   mat_err( row < 1,         "Error (mat_setz): Row index must be positive."             );
 //   mat_err( col < 1,         "Error (mat_setz): Column index must be positive."          );
 //   float complex *data = mat->data;
-//   ushort offset = (row-1) * (mat->cols) + (col-1);
+//   uint offset = (row-1) * (mat->cols) + (col-1);
 //   data += offset;
 //   *data = re + im *I;
 //   return;
@@ -378,15 +378,15 @@
 
 
 /*******************************************************************************
-* void mat_setrz ( matrixz* mat, ushort row, matrixz* vec )
+* void mat_setrz ( matrixz* mat, uint row, matrixz* vec )
 * Replaces a row of a complex matrix with the specified vector.
 *******************************************************************************/
-// void mat_setrz ( matrixz* mat, ushort row, matrixz* vec ) {
+// void mat_setrz ( matrixz* mat, uint row, matrixz* vec ) {
 //   mat_err( row > mat->rows,        "Error (mat_setrz): Row index exceeds matrix dimensions."           );
 //   mat_err( row < 1,                "Error (mat_setrz): Row index must be positive."                    );
 //   mat_err( vec->rows != 1,         "Error (mat_setrz): Input array must be a row vector."              );
 //   mat_err( mat->cols != vec->cols, "Error (mat_setrz): Input array and matrix must be the same width." );
-//   ushort i;
+//   uint i;
 //   for( i=1; i <= mat->cols; i++ )  mat_setz( mat, row, i, mat_getre( vec, 1, i ), mat_getim( vec, 1, i ) );
 //   return;
 // }
@@ -395,7 +395,7 @@
   mat_err( ( vec->rows !=1 ),           "Error (mat_setrz): Input array must be a row vector."              );
   mat_err( ( mat->cols != vec->cols ),  "Error (mat_setrz): Input array and matrix must be the same width." );
 
-  for( ushort i=0; i<mat->cols; i++ )  *( mat->data + (row-1) * (mat->cols) + i ) = *( vec->data+i );
+  for( uint i=0; i<mat->cols; i++ )  *( mat->data + (row-1) * (mat->cols) + i ) = *( vec->data+i );
 
   return;
 */
@@ -404,15 +404,15 @@
 
 
 /*******************************************************************************
-* void mat_setcz ( matrixz* mat, ushort col, matrixz* vec )
+* void mat_setcz ( matrixz* mat, uint col, matrixz* vec )
 * Replaces a column of a complex matrix with the specified vector.
 *******************************************************************************/
-// void mat_setcz ( matrixz* mat, ushort col, matrixz* vec ) {
+// void mat_setcz ( matrixz* mat, uint col, matrixz* vec ) {
 //   mat_err( col > mat->cols,        "Error (mat_setcz): Column index exceeds matrix dimensions."         );
 //   mat_err( col < 1,                "Error (mat_setcz): Column index must be positive."                  );
 //   mat_err( vec->cols != 1,         "Error (mat_setcz): Input array must be a column vector."            );
 //   mat_err( mat->rows != vec->rows, "Error (mat_setcz): Input array and matrix must be the same height." );
-//   ushort i;
+//   uint i;
 //   for( i=1; i <= mat->rows; i++ )  mat_setz( mat, i, col, mat_getre( vec, i, 1 ), mat_getim( vec, i, 1 ) );
 //   return;
 // }
@@ -421,7 +421,7 @@
   mat_err( ( vec->cols !=1 ),           "Error (mat_setcz): Input array must be a column vector."            );
   mat_err( ( mat->rows != vec->rows ),  "Error (mat_setcz): Input array and matrix must be the same height." );
 
-  for( ushort i=0; i<mat->rows; i++ )  *( mat->data + ( i * mat->cols ) + (col-1) ) = *( vec->data+i );
+  for( uint i=0; i<mat->rows; i++ )  *( mat->data + ( i * mat->cols ) + (col-1) ) = *( vec->data+i );
 
   return;
 */
